requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=0.15.0=pyhd3eb1b0_0
aiohttp=3.8.1=py38h7f8727e_0
aiosignal=1.2.0=pyhd3eb1b0_0
async-timeout=4.0.1=pyhd3eb1b0_0
attrs=21.4.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
blinker=1.4=py38h06a4308_0
bottleneck=1.3.4=py38hce1f21e_0
brotli=1.0.9=he6710b0_2
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f8727e_0
ca-certificates=2022.3.29=h06a4308_1
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2021.10.8=py38h06a4308_2
cffi=1.15.0=py38hd667e15_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.3=pyhd3eb1b0_0
cloudpickle=2.0.0=pypi_0
cpuonly=2.0=0
cryptography=3.4.8=py38hd23ed53_0
cycler=0.11.0=pyhd3eb1b0_0
dataclasses=0.8=pyh6d0b6a4_7
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py38h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
docker-pycreds=0.4.0=pypi_0
einops=0.4.1=pypi_0
entrypoints=0.3=py38_0
expat=2.4.4=h295c915_0
ffmpeg=4.3=hf484d3e_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
frozenlist=1.2.0=py38h7f8727e_0
giflib=5.2.1=h7b6447c_0
gitdb=4.0.9=pypi_0
gitpython=3.1.27=pypi_0
glib=2.69.1=h4ff587b_1
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=1.33.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.1=py_2
grpcio=1.42.0=py38hce63b2e_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
gym=0.9.1=pypi_0
icu=58.2=he6710b0_3
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.8.2=py38h06a4308_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.4.1=py38h06a4308_1
ipython=7.31.1=py38h06a4308_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.18.1=py38h06a4308_1
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
jupyter_client=7.1.2=pyhd3eb1b0_0
jupyter_core=4.9.1=py38h06a4308_0
kiwisolver=1.3.2=py38h295c915_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.19.1=h4ff587b_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.4=py38h06a4308_0
matplotlib=3.5.0=py38h06a4308_0
matplotlib-base=3.5.0=py38h3ed280b_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
multidict=5.2.0=py38h7f8727e_2
munkres=1.1.4=py_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.1=pyhd3eb1b0_0
nettle=3.7.3=hbbd107a_1
ninja=1.10.2=py38hd09550d_3
numexpr=2.8.1=py38h6abb31d_0
numpy=1.21.2=py38h20f2e39_0
numpy-base=1.21.2=py38h79a1101_0
oauthlib=3.1.1=pyhd3eb1b0_0
olefile=0.46=pyhd3eb1b0_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1n=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.2=py38h295c915_0
parso=0.8.3=pyhd3eb1b0_0
pathtools=0.1.2=pypi_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.4.0=py38h5aabda8_0
pip=21.2.4=py38h06a4308_0
promise=2.3=pypi_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.19.1=py38h295c915_0
psutil=5.9.0=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pyglet=1.5.21=pypi_0
pygments=2.11.2=pyhd3eb1b0_0
pyjwt=2.1.0=py38h06a4308_0
pyopenssl=21.0.0=pyhd3eb1b0_1
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py38h05f1152_4
pysocks=1.7.1=py38h06a4308_0
python=3.8.12=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.8.0=py3.8_cpu_0
pytorch-mutex=1.0=cpu
pytz=2021.3=pyhd3eb1b0_0
pyyaml=6.0=py38h7f8727e_1
pyzmq=22.3.0=py38h295c915_2
qt=5.9.7=h5867ecd_1
readline=8.1.2=h7f8727e_1
requests=2.27.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-learn=1.0.2=py38h51133e4_1
scipy=1.7.3=py38hc147768_0
seaborn=0.11.2=pyhd3eb1b0_0
sentry-sdk=1.5.8=pypi_0
setproctitle=1.2.2=pypi_0
setuptools=58.0.4=py38h06a4308_0
shortuuid=1.0.8=pypi_0
signatory=1.2.6.1.8.0=pypi_0
sip=4.19.13=py38h295c915_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
sqlite=3.37.2=hc218d9a_0
tensorboard=2.6.0=py_1
tensorboard-data-server=0.6.0=py38hca6d32c_0
tensorboard-plugin-wit=1.6.0=py_0
tensorboardx=2.2=pyhd3eb1b0_0
termcolor=1.1.0=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.11=h1ccaba5_0
torchaudio=0.8.0=py38
torchvision=0.9.0=py38_cpu
tornado=6.1=py38h27cfd23_0
tqdm=4.62.3=pyhd3eb1b0_1
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=3.10.0.2=hd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.26.8=pyhd3eb1b0_0
wandb=0.12.11=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
werkzeug=2.0.2=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
yarl=1.6.3=py38h27cfd23_0
yaspin=2.1.0=pypi_0
zeromq=4.3.4=h2531618_0
zipp=3.7.0=pyhd3eb1b0_0
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0
